fix can't delete if not upload image

+++ controllers/pr-controller.js
@@ -1,25 +1,25 @@
 const { Pr_monitoring } = require('../models')
 const { Section_tab } = require('../models')
-const  fs = require( 'fs')
+const fs = require('fs')
 
 module.exports = {
     index: async (req, res) => {
         try {
             const pr = await Pr_monitoring.findAll();
             res.status(200).json(pr);
-        }catch(e){
+        } catch (e) {
             res.status(500).json(e)
         }
     },
     store: async (req, res) => {
         try {
             let itemDescImgPath = ''; // Initialize with an empty string
-    
+
             // Check if an image file was uploaded
             if (req.file) {
-                itemDescImgPath = `/images/${req.file.filename}`;
+                itemDescImgPath = `/uploads/${req.file.filename}`;
             }
-    
+
             const document = await Pr_monitoring.create({
                 req_date: req.body.req_date,
                 item_desc_img: itemDescImgPath,
@@ -37,7 +37,7 @@ module.exports = {
                 bidding: req.body.bidding,
                 keterangan: req.body.keterangan,
             });
-    
+
             res.status(200).json(document);
         } catch (error) {
             console.error(error)
@@ -46,31 +46,34 @@ module.exports = {
     },
     delete: async (req, res) => {
         try {
-          const document = await Pr_monitoring.findOne({
-            where: {
-              id: req.params.id,
-            },
-          });
-          const check = fs.existsSync(`public/${document.item_desc_img}`);
-          if (check) {
-            fs.unlinkSync(`public/${document.item_desc_img}`);
-          }
-          await document.destroy();
-          res.status(200).json({ message: "Data deleted" });
+            const document = await Pr_monitoring.findOne({
+                where: {
+                    id: req.params.id,
+                },
+            });
+            const check = fs.existsSync(`public/${document.item_desc_img}`);
+
+            if (check) {
+                if (document.item_desc_img != "") {
+                    fs.unlinkSync(`public/${document.item_desc_img}`);
+                }
+            }
+            await document.destroy();
+            res.status(200).json({ message: "Data deleted" });
         } catch (error) {
             console.log(error);
-          res.status(500).json(error.message);
+            res.status(500).json(error.message);
         }
-      },
-    
+    },
+
     section: async (req, res) => {
         try {
             const sec = await Section_tab.findAll();
             res.status(200).json(sec);
-        }catch(e){
+        } catch (e) {
             res.status(500).json(e)
         }
     },
-    
-    
+
+
 }
